+++ pkgs/http/CHANGELOG.md
@@ -3,6 +3,8 @@
 * **Breaking** Change the behavior of `Request.body` so that a charset
   parameter is only added for text and XML media types. This brings the
   behavior of `package:http` in line with RFC-8259.
+* On the web, fix cancellations for `StreamSubscription`s of response bodies
+  waiting for the next chunk.
 * Export `MediaType` from `package:http_parser`.
 * Added a section on testing to `README.md`.